backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.31.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.23.8
pdfminer.six==20231228
pypdf==4.0.1
pytesseract==0.3.10
Pillow==10.2.0